New inputs to the tool and adoption of profile factory
The tool now gets as inputs: The name of the folder containing the CIM model (divided in CIM profiles) and the given name for the target model.
Main class has a new set up function to load the internal profile factory. Creates the necessary profile classes according to the available profile files